+++ Autocad_ConcerteList/Src/ConcreteDB/DataObjects/SerieDb.cs
@@ -8,7 +8,7 @@ namespace Autocad_ConcerteList.Src.ConcreteDB.DataObjects
 {
     public class SerieDbo
     {
-        public decimal SeriesId { get; set; }        
+        public int SeriesId { get; set; }        
         public string Name { get; set; }
 
         public override string ToString ()
